+++ Rewards/Classes/Account/Account.swift
@@ -3,7 +3,6 @@
  * MIT license. See LICENSE file in the root directory.
  */
 
-
 import Foundation
 import SwiftUI
