+++ Rewards/Classes/Home/HomeAccountGrid.swift
@@ -7,7 +7,7 @@ import SwiftUI
 
 public struct HomeAccountGrid: View{
     
-    @Binding var accountsList
+    @Binding var accountsList: [Account]
 
     private var gridItemLayout = Array(repeating: GridItem(.flexible(), spacing: 0), count: 3)
     
@@ -23,7 +23,7 @@ public struct HomeAccountGrid: View{
                     ForEach(accountsList, id: \.accountCommon.name){ acc in
                         VStack(
                           alignment: /*@START_MENU_TOKEN@*/.center/*@END_MENU_TOKEN@*/) {
-                            SheetHomeCarouselIcon(provider: acc, width: 77, height: 77)
+                            AccountIcon(provider: acc, width: 77, height: 77)
                               .padding(.bottom, -15)
                         Text("Add")
                           .font(SpaceGrotesk.medium(size: 12))
