+++ Rewards/Classes/Home/HomeCarousel.swift
@@ -15,7 +15,7 @@ public struct HomeCarousel : View {
             HStack(spacing: 20) {
                 ForEach(accountsList, id: \.accountCommon.name){ acc in
                     VStack(alignment: /*@START_MENU_TOKEN@*/.center/*@END_MENU_TOKEN@*/) {
-                        HomeAccountIcon(provider: acc, width: 77, height: 77)
+                        AccountIcon(provider: acc, width: 77, height: 77)
                         Text("Add")
                           .font(SpaceGrotesk.medium(size: 12))
                           .foregroundColor(.tikiDarkGray)
