+++ Rewards/Classes/Retailer/RetailerOffer.swift
@@ -1,0 +1,12 @@
+/*
+ * Copyright (c) TIKI Inc.
+ * MIT license. See LICENSE file in the root directory.
+ */
+
+import Foundation
+
+struct RetailerOffer{
+    let account: Account
+    let title: String
+    let url: URL
+}
